requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
websockets>=12.0
crewai>=0.30.1
langchain-openai>=0.1.7
langchain-anthropic>=0.1.15
python-dotenv>=1.0.1
pydantic>=2.7.1
httpx>=0.27.0
